refactor(mobile): implement test directory organization structure
- Create dedicated tests/ directory with organized subdirectories:
  - tests/integration/ - Cross-component testing
  - tests/e2e/ - End-to-end scenarios
  - tests/performance/ - Performance benchmarks
  - tests/acceptance/ - Business requirement validation
- Move integration tests to dedicated directory:
  - useAuthenticationIntegration.test.ts (auth flow testing)
  - AuthenticationOrchestrator.test.ts (orchestrator integration)
  - StoreContext.test.tsx (store context integration)
- Update all import paths in moved tests to reflect new directory structure
- Add comprehensive README files documenting test categories and guidelines
- Maintain Jest config compatibility with existing testMatch patterns

This completes Step 2 of the testing refactoring plan, establishing clear separation
between unit tests (co-located) and integration tests (dedicated directory).
