+++ apps/mobile/tests/acceptance/README.md
@@ -1,0 +1,30 @@
+# Acceptance Tests
+
+This directory contains tests that validate business requirements and user acceptance criteria.
+
+## Test Categories
+
+- **Business Logic Validation**: Verify core business rules are enforced
+- **User Story Tests**: Test acceptance criteria from user stories
+- **Regulatory Compliance**: Ensure compliance requirements are met
+- **Feature Completeness**: Validate all feature requirements
+
+## Guidelines
+
+- Write tests in business language (Given-When-Then format)
+- Focus on user value and business outcomes
+- Test business rules and constraints
+- Include edge cases and boundary conditions
+
+## Example Test Structure
+
+```typescript
+// authentication.test.ts
+describe('Authentication Business Requirements', () => {
+  it('should enforce wallet signature requirements per security policy', async () => {
+    // Given a user attempts authentication
+    // When they provide an invalid signature
+    // Then the system should reject the attempt
+  })
+})
+```
