+++ apps/mobile/tests/performance/README.md
@@ -1,0 +1,28 @@
+# Performance Tests
+
+This directory contains tests focused on performance benchmarks and optimization validation.
+
+## Test Categories
+
+- **MobX Store Performance**: Measure store update times and reaction efficiency
+- **Component Render Performance**: Test component mounting and update performance
+- **Memory Usage Tests**: Monitor memory consumption patterns
+- **Network Performance**: Test API call efficiency and caching
+
+## Guidelines
+
+- Use performance measurement utilities (React DevTools Profiler, etc.)
+- Set performance budgets and thresholds
+- Test both initial load and ongoing usage patterns
+- Include tests for memory leaks and cleanup
+
+## Example Test Structure
+
+```typescript
+// store-updates.test.ts
+describe('MobX Store Performance', () => {
+  it('should update authentication store within performance budget', async () => {
+    // Measure store update performance
+  })
+})
+```
